find-fixes: Fix incremental updates
Incremental updates of a fixes file is currently broken since the script
assumes a linear commit history which is not the case. Fix that by storing
the HEAD commit at the top of the fixes file so that we know where to start
looking from the next time around.

While at it, make the usage more verbose and actually useful.

Signed-off-by: Juerg Haefliger <juergh@gmail.com>